+++ gkraken/di.py
@@ -20,7 +20,7 @@ import logging
 from typing import Optional, NewType
 from gi.repository import Gtk
 from injector import Module, provider, singleton, Injector
-from liquidctl.cli import find_all_supported_devices
+from liquidctl.driver import find_liquidctl_devices
 from liquidctl.driver.kraken_two import KrakenTwoDriver
 from peewee import SqliteDatabase
 from rx.disposable import CompositeDisposable
@@ -85,7 +85,7 @@ class ProviderModule(Module):
     @provider
     def provide_kraken_two_driver(self) -> Optional[KrakenTwoDriver]:
         _LOG.debug("provide KrakenTwoDriver")
-        return next((dev for dev in find_all_supported_devices() if isinstance(dev, KrakenTwoDriver)), None)
+        return next((dev for dev in find_liquidctl_devices() if isinstance(dev, KrakenTwoDriver)), None)
 
     @singleton
     @provider